Add basic scene character and basic gamestate logic.
In this commit GameState, pukeko and the ball were added. The game state holds information about the game like the hoop and ball.  There is currently a test script added to the pukeko that pushes the ball and jumps toward the hoop the pukeko is on. Next update we need to feed all of that into the NEAT addon that was added as well